Simple Reaction Time Task raw response data

Simple reaction time study (Simple Reaction Time Task), which is included in JATOS examples. The example study was uploaded to MindProbe server and run four times. See process_results on how to turn the raw data into processed values.

Usage

demo_srt

Format

demo_srt

A list with two items:

meta

httr2_response as returned by get_metadata

results

httr2_response as returned by get_results, with body referring to the bundled srtdemo.zip file, which you can copy into current working directory with prepare_demo

Source

https://github.com/visionlabels/jatosR

Details

This is the raw response data with two httr2_response values, as used by prepare_demo().